requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.3
PySocks==1.7.1
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
selenium==4.22.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
telebot==0.0.5
trio==0.25.1
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
watchfiles==0.22.0
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
